refactor: juntar CardHorizontal e CardVertical em um mesmo arquivo
