+++ src/components/CardProduct/styles.sass
@@ -1,4 +1,5 @@
-@use '../../../styles/variables' as *
+@use '../../styles/variables' as *
+@use '../../styles/mixins' as *
 
 .cart-item
   background-color: $light-color
@@ -178,3 +179,128 @@
     top: 0.5rem
     right: 0.5rem
     font-size: 1.25rem
+
+
+.card-product
+    &-home, &-list, &-favorite
+        height: 450px
+        display: flex
+        flex-direction: column
+        justify-content: center
+        align-items: center
+        border: 1px solid $gray-text-color
+        border-radius: 8px
+        transition: .4s, box-shadow .4s
+        padding: 1rem
+        position: relative
+
+        &:hover
+            border-color: $primary-color
+            transform: translateY(-5px)
+            box-shadow: 0 10px 25px rgba(0, 0, 0, 0.15)
+
+    &-home
+        width: 20%
+        position: relative
+
+        @include mobile
+            width: 90%
+            margin: 1rem
+
+            .card-product-image 
+                height: 40%
+
+                img
+                    width: 50%
+
+    &-favorite
+        background-color: $light-color
+
+        @include mobile
+            width: 90%
+            height: 380px
+
+            .card-product-image 
+                height: 40%
+                
+                img
+                    width: 50%
+
+.btn-favorite
+    width: 55px
+    height: 55px
+    background: none
+    border: none
+    color: $primary-color
+    cursor: pointer
+    transition: transform 0.4s
+    position: absolute
+    top: 0
+    left: 0
+
+    &:hover
+        transform: scale(1.1)
+
+.label
+    font-size: 0.75rem
+    font-weight: 700
+    text-transform: uppercase
+    color: $primary-color
+    border-radius: 4px
+    padding: 0.35rem 0.75rem
+    z-index: 1
+    position: absolute
+    top: 1rem
+    right: 1rem
+
+.promo
+    background-color: $primary-color
+    color: $light-color
+
+.card-product
+    &-image
+        width: 100%
+        height: 200px
+        display: flex
+        align-items: center
+        justify-content: center
+        padding: 1rem
+        background-color: $light-color
+
+        img
+            max-width: 100%
+            max-height: 100%
+            object-fit: contain
+
+    &-info
+        padding: 1.5rem
+        text-align: center
+
+        .product
+            &-category, &-name, &-price
+                margin-bottom: 1rem
+
+            &-category
+                display: block
+                font-size: 0.75rem
+                font-weight: 600
+                text-transform: uppercase
+                letter-spacing: 0.5px
+                color: $gray-text-color
+
+            &-name
+                min-height: 2.8rem
+                display: -webkit-box
+                -webkit-line-clamp: 2
+                -webkit-box-orient: vertical
+                font-size: 1rem
+                font-weight: 600
+                text-align: center
+                line-height: 1.4
+                color: $dark-color
+                overflow: hidden
+
+            &-price 
+                color: $primary-color
+                font-size: 1.5rem
+                font-weight: 700
