+++ src/pages/shoppingCar/index.jsx
@@ -2,7 +2,7 @@ import { useContext } from 'react';
 
 import { ShoppingCartContext } from '../../context/shoppingCart';
 
-import CardHorizontal from '../../components/CardProduct/CardHorizontal';
+import { CardHorizontal } from '../../components/CardProduct';
 import AsideShoppingCar from '../../components/Aside/AsideShoppingCar';
 import Footer from '../../layout/Footer';
